chore: add generate rollback script
