--- scripts/rollback.sh
@@ -1,39 +1,0 @@
-#!/bin/bash
-# timestamp: 1600853087221
-
-npm dist-tag add @midwayjs/bootstrap@2.2.0 latest
-tnpm dist-tag add @midwayjs/bootstrap@2.2.0 latest
-npm dist-tag add @midwayjs/core@2.1.4 latest
-tnpm dist-tag add @midwayjs/core@2.1.4 latest
-npm dist-tag add @midwayjs/decorator@1.20.3 latest
-tnpm dist-tag add @midwayjs/decorator@1.20.3 latest
-npm dist-tag add @midwayjs/faas@1.2.1 latest
-tnpm dist-tag add @midwayjs/faas@1.2.1 latest
-npm dist-tag add midway-schedule@1.20.3 latest
-tnpm dist-tag add midway-schedule@1.20.3 latest
-npm dist-tag add midway@1.20.3 latest
-tnpm dist-tag add midway@1.20.3 latest
-npm dist-tag add @midwayjs/mock@2.2.0 latest
-tnpm dist-tag add @midwayjs/mock@2.2.0 latest
-npm dist-tag add @midwayjs/socketio@2.2.0 latest
-tnpm dist-tag add @midwayjs/socketio@2.2.0 latest
-npm dist-tag add @midwayjs/express@2.2.0 latest
-tnpm dist-tag add @midwayjs/express@2.2.0 latest
-npm dist-tag add @midwayjs/koa@2.2.0 latest
-tnpm dist-tag add @midwayjs/koa@2.2.0 latest
-npm dist-tag add @midwayjs/web@2.2.0 latest
-tnpm dist-tag add @midwayjs/web@2.2.0 latest
-
-# Changes:
-
-#  - @midwayjs/bootstrap: 2.2.0 => 3.2.9
-#  - @midwayjs/core: 2.1.4 => 3.2.9
-#  - @midwayjs/decorator: 1.20.3 => 3.2.9
-#  - @midwayjs/faas: 1.2.1 => 3.2.9
-#  - midway-schedule: 1.20.3 => 3.2.9
-#  - midway: 1.20.3 => 3.2.9
-#  - @midwayjs/mock: 2.2.0 => 3.2.9
-#  - @midwayjs/socketio: 2.2.0 => 3.2.9
-#  - @midwayjs/express: 2.2.0 => 3.2.9
-#  - @midwayjs/koa: 2.2.0 => 3.2.9
-#  - @midwayjs/web: 2.2.0 => 3.2.9
