+++ scripts/rollback/rollback-2.8.0.sh
@@ -1,0 +1,51 @@
+#!/bin/bash
+
+# Changes:
+# - @midwayjs/bootstrap: 2.7.7 => 2.8.0
+# - @midwayjs/core: 2.7.7 => 2.8.0
+# - @midwayjs/decorator: 2.7.7 => 2.8.0
+# - @midwayjs/faas: 2.7.7 => 2.8.0
+# - @midwayjs/grpc: 2.7.7 => 2.8.0
+# - @midwayjs/logger: 2.7.7 => 2.8.0
+# - midway-schedule: 2.7.7 => 2.8.0
+# - midway: 2.7.7 => 2.8.0
+# - @midwayjs/mock: 2.7.7 => 2.8.0
+# - @midwayjs/rabbitmq: 2.7.7 => 2.8.0
+# - @midwayjs/socketio: 2.7.7 => 2.8.0
+# - @midwayjs/typegoose: 2.7.7 => 2.8.0
+# - @midwayjs/express: 2.7.7 => 2.8.0
+# - @midwayjs/koa: 2.7.7 => 2.8.0
+# - @midwayjs/web: 2.7.7 => 2.8.0
+
+npm dist-tag add @midwayjs/bootstrap@2.8.0 beta
+npm dist-tag add @midwayjs/bootstrap@2.7.5 latest
+npm dist-tag add @midwayjs/core@2.8.0 beta
+npm dist-tag add @midwayjs/core@2.7.5 latest
+npm dist-tag add @midwayjs/decorator@2.8.0 beta
+npm dist-tag add @midwayjs/decorator@2.7.5 latest
+npm dist-tag add @midwayjs/faas@2.8.0 beta
+npm dist-tag add @midwayjs/faas@2.7.7 latest
+npm dist-tag add @midwayjs/grpc@2.8.0 beta
+npm dist-tag add @midwayjs/grpc@2.7.7 latest
+npm dist-tag add @midwayjs/logger@2.8.0 beta
+npm dist-tag add @midwayjs/logger@2.7.5 latest
+npm dist-tag add midway-schedule@2.8.0 beta
+npm dist-tag add midway-schedule@2.7.7 latest
+npm dist-tag add midway@2.8.0 beta
+npm dist-tag add midway@2.7.7 latest
+npm dist-tag add @midwayjs/mock@2.8.0 beta
+npm dist-tag add @midwayjs/mock@2.7.7 latest
+npm dist-tag add @midwayjs/rabbitmq@2.8.0 beta
+npm dist-tag add @midwayjs/rabbitmq@2.7.7 latest
+npm dist-tag add @midwayjs/socketio@2.8.0 beta
+npm dist-tag add @midwayjs/socketio@2.7.7 latest
+npm dist-tag add @midwayjs/typegoose@2.8.0 beta
+npm dist-tag add @midwayjs/typegoose@2.7.7 latest
+npm dist-tag add @midwayjs/typegoose@2.8.0 beta
+npm dist-tag add @midwayjs/typegoose@2.7.7 latest
+npm dist-tag add @midwayjs/express@2.8.0 beta
+npm dist-tag add @midwayjs/express@2.7.7 latest
+npm dist-tag add @midwayjs/koa@2.8.0 beta
+npm dist-tag add @midwayjs/koa@2.7.7 latest
+npm dist-tag add @midwayjs/web@2.8.0 beta
+npm dist-tag add @midwayjs/web@2.7.7 latest
